♻️ refactor: 이미지 캐시 키 유틸 공통화
중복된 이미지 캐시 키 생성 로직을 공용 유틸로 추출하고 불필요한 non-null assertion을 제거합니다.

리뷰 반영 #3407184349 #3407187252 #3407187256 #3407187259 #3407187261 #3407187262 #3407187266 #3407187271
